Fix compilation on Windows on the 2.4 branch after revert of aggresive garbage collection when close to OOM patch.
TBR=antonm@chromium.org
Review URL: http://codereview.chromium.org/3966003

git-svn-id: http://v8.googlecode.com/svn/branches/2.4@5682 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.cc
@@ -691,7 +691,7 @@ void Heap::PerformGarbageCollection(GarbageCollector collector,
 
     UpdateSurvivalRateTrend(start_new_space_size);
 
-    int old_gen_size = PromotedSpaceSize();
+    intptr_t old_gen_size = PromotedSpaceSize();
     old_gen_promotion_limit_ =
         old_gen_size + Max(kMinimumPromotionLimit, old_gen_size / 3);
     old_gen_allocation_limit_ =
